requirements.txt
Django==5.1.6
django_environ==0.12.0
django_guardian==2.4.0
django_import_export==4.3.7
django_tables2==2.7.5
openpyxl==3.1.5
Pillow==11.1.0
reportlab==4.3.1
